third_party/blink/web_tests/platform/mac-mac15-arm64/external/wpt/html/canvas/offscreen/manual/the-offscreen-canvas/offscreencanvas.getcontext-expected.txt

This is a testharness.js-based test.
[FAIL] Test that getContext with supported string returns correct results
  assert_true: expected true got false
[FAIL] Test that getContext twice with different context type returns null the second time
  assert_equals: expected null but got object "[object OffscreenCanvasRenderingContext2D]"
[FAIL] Test that webglcontext.canvas should return the original OffscreenCanvas
  Cannot read properties of null (reading 'canvas')
Harness: the test ran to completion.
